requirements.txt
selenium==3.141.0
marshmallow==2.21.0
psycopg2-binary==2.8.6
Flask==1.1.4